\name{return_codeCLP}
\alias{return_codeCLP}

\title{
  Translates a COIN-OR Clp Return Code into a Human
  Readable String
}

\description{
  Translates a COIN-OR Clp return code into a human
  readable string.
}

\usage{
  return_codeCLP(code)
}

\arguments{
  \item{code}{
    Return code from COIN-OR Clp.
  }
}

\value{
  A character string associated with the COIN-OR Clp
  return code.
}

\references{
   The COIN-OR Clp home page at
   \url{https://projects.coin-or.org/Clp}
}

\author{
  Gabriel Gelius-Dietrich <geliudie@uni-duesseldorf.de>

  Maintainer: Mayo Roettger <mayo.roettger@hhu.de>
}


\keyword{ optimize }

