\name{status_codeCLP}
\alias{status_codeCLP}

\title{
  Translates a COIN-OR Clp Status Value into a Human
  Readable String
}

\description{
  Translates a COIN-OR Clp status value into a human
  readable string.
}

\usage{
  status_codeCLP(code)
}

\arguments{
  \item{code}{
    Status code from COIN-OR Clp.
  }
}

\value{
  A character string associated with the COIN-OR Clp
  status code.
}

\references{
   The COIN-OR Clp home page at
   \url{http://www.coin-or.org/projects/Clp.xml}
}

\author{
  Gabriel Gelius-Dietrich <geliudie@uni-duesseldorf.de>

  Maintainer: Claus Jonathan Fritzemeier <clausjonathan.fritzemeier@uni-duesseldorf.de>
}


\keyword{ optimize }

